requirements.txt
asgiref==3.6.0
attrs==22.2.0
colorama==0.4.6
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.0
Pillow==9.5.0
pluggy==1.0.0
PyJWT==2.6.0
pytest==7.2.2
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
